+++ interpolation/gsl_interp.h
@@ -7,9 +7,9 @@
 
 /* evaluation accelerator */
 typedef struct {
-  unsigned long  cache;        /* cache of index   */
-  unsigned long  miss_count;   /* keep statistics  */
-  unsigned long  hit_count;
+  size_t  cache;        /* cache of index   */
+  size_t  miss_count;   /* keep statistics  */
+  size_t  hit_count;
 }
 gsl_interp_accel;
 
@@ -21,7 +21,7 @@ struct _gsl_interp_obj_struct {
   void    (*free)        (struct _gsl_interp_obj_struct *);
   double  xmin;
   double  xmax;
-  int     size;
+  size_t     size;
 };
 typedef  struct _gsl_interp_obj_struct  gsl_interp_obj;
 
@@ -29,7 +29,7 @@ typedef  struct _gsl_interp_obj_struct  gsl_interp_obj;
 /* interpolation object factory */
 typedef struct {
   const char * name;
-  gsl_interp_obj *  (*create) (const double x_array[], const double y_array[], int size);
+  gsl_interp_obj *  (*create) (const double x_array[], const double y_array[], size_t size);
 }
 gsl_interp_factory;
 
@@ -45,8 +45,8 @@ extern const gsl_interp_factory   gsl_interp_factory_akima_periodic;
 gsl_interp_accel *
 gsl_interp_accel_new(void);
 
-unsigned long
-gsl_interp_accel_find(gsl_interp_accel * a, const double x_array[], unsigned long size, double x);
+size_t
+gsl_interp_accel_find(gsl_interp_accel * a, const double x_array[], size_t size, double x);
 
 void
 gsl_interp_accel_free(gsl_interp_accel * a);
@@ -98,10 +98,10 @@ gsl_interp_obj_free(gsl_interp_obj * interp_obj);
 #include "bsearch.h"
 extern
 inline
-unsigned long
-gsl_interp_accel_find(gsl_interp_accel * a, const double xa[], unsigned long len, double x)
+size_t
+gsl_interp_accel_find(gsl_interp_accel * a, const double xa[], size_t len, double x)
 {
-  unsigned long x_index = a->cache;
+  size_t x_index = a->cache;
  
   if(x < xa[x_index]) {
     a->miss_count++;
